+++ test/HelperContract.sol
@@ -15,7 +15,7 @@ abstract contract HelperContract is Test, AuthorizationModule{
     CCIPSender CCIPSENDER_CONTRACT;
     // EOA to perform tests
     address constant ZERO_ADDRESS = address(0);
-    address constant DEFAULT_ADMIN_ADDRESS = address(1);
+    //address constant DEFAULT_ADMIN_ADDRESS = address(1);
     // Operator
     address constant DEBT_VAULT_OPERATOR_ADDRESS = address(2);
     address constant DEBT_VAULT_DEPOSIT_OPERATOR_ADDRESS = address(3);
@@ -28,11 +28,12 @@ abstract contract HelperContract is Test, AuthorizationModule{
     address constant TOKEN_PAYMENT_ADMIN = address(8);
     address constant CCIPSENDER_ADMIN = address(0x9);
     address constant RECEIVER_ADDRESS = address(0xA);
+    address constant SENDER_ADDRESS = address(0xB);
 
     uint64 AVALANCHE_SELECTOR = 6433500567565415381;
     IERC20 AVALANCHE_USDC = IERC20(0xB97EF9Ef8734C71904D8002F8b6Bc66Dd9c48a6E);
+    
     // role string
-
     string constant DEFAULT_ADMIN_ROLE_HASH =
         "0x0000000000000000000000000000000000000000000000000000000000000000";